==> 2-Traitement donnees Solaire_Vents100m_PROC.ipynb <==
"""
Created on Wed Feb  5 18:27:48 2025
@author: Thierry ALLEM
"""

# TRAITEMENT DES DONNEES SUR LE RAYONNEMENT SOLAIRE ET LA VITESSE DU VENTS A 100M (pour l'étude sur la production des éoliennes)
# Approximations sur la base d'une base de données de relevés tri-horaires

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Importation du fichier des relevés de températures
# Lecture du fichier CSV
df_vent_100m_solaire_3h = pd.read_csv("solaire_vent_regions.csv", sep=';', encoding='utf-8')
df_vent_100m_solaire_3h.head()

df_vent_100m_solaire_3h.info()

# Conversion de la colonne 'date_heure' en datetime
df_vent_100m_solaire_3h['date_heure'] = pd.to_datetime(df_vent_100m_solaire_3h['date_heure'], errors='coerce', utc=True)
df_vent_100m_solaire_3h["date_heure"] = df_vent_100m_solaire_3h["date_heure"].dt.tz_localize(None)


# Extraction de l'année
df_vent_100m_solaire_3h['annee'] = df_vent_100m_solaire_3h['date_heure'].dt.year

# Ajout d'une colonne 'date' avec le format 'yyyy-mm-dd'
df_vent_100m_solaire_3h['date'] = df_vent_100m_solaire_3h['date_heure'].dt.strftime('%Y-%m-%d')

# Ajout d'une colonne 'heure' à l'heure française
df_vent_100m_solaire_3h['heure'] = df_vent_100m_solaire_3h['date_heure'].dt.strftime('%H:%M')
df_vent_100m_solaire_3h.info()
df_vent_100m_solaire_3h['annee'].unique()
df_vent_100m_solaire_3h['heure'].unique()

# Filtrage des années pour limiter le dataset de 2013 à 2022 inclus
df_vent_100m_solaire_3h = df_vent_100m_solaire_3h[(df_vent_100m_solaire_3h["annee"] >= 2013) & (df_vent_100m_solaire_3h["annee"] <= 2022)]

# Liste des régions
df_vent_100m_solaire_3h['region'].unique()

# Suppression des lignes de régions non concernées par cette étude
regions_a_exclure = [np.nan, 'Corse']
df_vent_100m_solaire_3h = df_vent_100m_solaire_3h[~df_vent_100m_solaire_3h['region'].isin(regions_a_exclure)]
df_vent_100m_solaire_3h['region'].unique()

# Renommage des valeurs de 'region'
renommage_regions = {
    'Occitanie': 'OCCITANIE',
    'Normandie': 'NORMANDIE',
    'Bretagne': 'BRETAGNE',
    'Provence-Alpes-Côte d\'Azur': 'PROVENCE ALPES COTE D AZUR',
    'Hauts-de-France': 'HAUTS DE FRANCE',
    'Île-de-France': 'ILE DE FRANCE',
    'Grand Est': 'GRAND EST',
    'Nouvelle-Aquitaine': 'NOUVELLE AQUITAINE',
    'Bourgogne-Franche-Comté': 'BOURGOGNE FRANCHE COMTE',
    'Pays de la Loire': 'PAYS DE LA LOIRE',
    'Auvergne-Rhône-Alpes': 'AUVERGNE RHONE ALPES',
    'Centre-Val de Loire': 'CENTRE VAL DE LOIRE'
}
df_vent_100m_solaire_3h['region'] = df_vent_100m_solaire_3h['region'].replace(renommage_regions)
df_vent_100m_solaire_3h['region'].unique()

# INTERPOLATION DES DONNEES SUR LA PERIODE 2013 -2022 PAR PAS DE 30 MINUTES

# Création de la plage temporelle de 2013-01-01 à 2022-12-31 avec un pas de 30 minutes
date_range = pd.date_range("2013-01-01 00:00", "2022-12-31 23:30", freq="30min")

# Liste des régions 
regions = df_vent_100m_solaire_3h['region'].unique()

# Construction du DataFrame initial 
df_base = pd.DataFrame({"date_heure": np.tile(date_range, len(regions)),
                        "region": np.repeat(regions, len(date_range))})

# Fusion des données initiales
merged_df = pd.merge(df_base, df_vent_100m_solaire_3h,
                      on=["date_heure", "region"],
                      how="left")

# Traitement des années 2013-2015 
merged_df['annee'] = merged_df['date_heure'].dt.year
mask_zeros = merged_df['annee'].isin([2013, 2014, 2015])
merged_df.loc[mask_zeros, ['Vitesse du vent à 100m (m/s)', 'Rayonnement solaire global (W/m2)']] = 0.0

# Interpolation des valeurs par région
merged_df.sort_values(by=['region', 'date_heure'], inplace=True)
merged_df[['Vitesse du vent à 100m (m/s)', 'Rayonnement solaire global (W/m2)']] = (
    merged_df.groupby('region', group_keys=False)[['Vitesse du vent à 100m (m/s)', 'Rayonnement solaire global (W/m2)']]
    .apply(lambda group: group.interpolate(method='linear'))
)

# Nettoyage des colonnes inutiles
df_vent_100m_solaire_30min = merged_df[['date_heure', 'region', 'Vitesse du vent à 100m (m/s)', 'Rayonnement solaire global (W/m2)']]
df_vent_100m_solaire_30min.info()

# Renommage de noms de colonnes

colonnes_a_renommer = {'Vitesse du vent à 100m (m/s)':'vitesse_vent_100m_m_par_s',
                       'Rayonnement solaire global (W/m2)':'rayonnement_solaire_global_W_par_m2'}
df_vent_100m_solaire_30min=df_vent_100m_solaire_30min.rename(columns=colonnes_a_renommer)
df_vent_100m_solaire_30min.info()

df_vent_100m_solaire_30min.head(20)

# Tri et sauvegarde des résultats
df_vent_100m_solaire_30min = df_vent_100m_solaire_30min.sort_values(['region', 'date_heure'])
df_vent_100m_solaire_30min.to_csv("df_vent_100m_solaire_30min.csv", sep=';',index=False)
