==> 3-SYNOP_ECO2Mix fusion_PROC.ipynb <==
"""
Created on Wed Oct  9 18:51:52 2024

@author: Thierry ALLEM
"""


# ***************************************** CALCULS DES CONSOMMATIONS CUMULEE - AJOUT DES DONNEES METEOROLOGIQUES ET D'INDICATEURS ****

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Importation du fichier eco2mix avec consommations
# Lecture du fichier CSV
df_eco_2_cumuls_delta = pd.read_csv("df_eco2_powerbi.csv", sep=';', encoding='latin-1')
df_eco_2_cumuls_delta.head()

df_eco_2_cumuls_delta.info()

# Suppression de colonnes non utiles dans cette étude
colonnes_a_supprimer = [
    'Unnamed: 0', 'region_abr','mois_nom','tch_6f_rte','tch_consommation_rte',
    'tch_thermique_rte','tch_nucleaire_rte','tch_eolien_rte','tch_solaire_rte','tch_hydraulique_rte','tch_bioenergies_rte','cap_prod_max_thermique',
    'thermique_p_disp','cap_prod_max_nucleaire',
    'nucleaire_p_disp','eolien_p_disp','solaire_p_disp','ind_prod_hydraulique','hydraulique_p_disp','cap_prod_max_bioenergies',
    'cap_prod_max_hydraulique','bioenergies_p_disp','region_p_max','cmax_thermique_rte','cmax_nucleaire_rte',
    'cmax_eolien_rte','cmax_solaire_rte','cmax_hydraulique_rte','cmax_bioenergies_rte','jour_ferie', 'saison','clst_tch_6f',
    'clst_tch_consommation','clst_tch_consommation_isol','date_heure_modifiee']
df_eco_2_cumuls_delta.drop(columns=colonnes_a_supprimer, inplace=True)
df_eco_2_cumuls_delta.info()

# Conversion de la colonne 'date_heure' en datetime
df_eco_2_cumuls_delta['date_heure'] = pd.to_datetime(df_eco_2_cumuls_delta['date_heure'])

# Ajour d'une colonne de nom de mois, en français

# Dictionnaire de correspondance des noms de mois en français (avec majuscule)
mois_mapping = {
    1: 'Janvier', 2: 'Février', 3: 'Mars', 4: 'Avril', 5: 'Mai', 6: 'Juin',
    7: 'Juillet', 8: 'Août', 9: 'Septembre', 10: 'Octobre', 11: 'Novembre', 12: 'Décembre'
}

# Ajout de la colonne 'mois_nom'
df_eco_2_cumuls_delta['mois_nom'] = df_eco_2_cumuls_delta['date_heure'].dt.month.map(mois_mapping)

# Extraction de la date uniquement (sans les heures)
df_eco_2_cumuls_delta['date'] = df_eco_2_cumuls_delta['date_heure'].dt.strftime('%Y-%m-%d')

# Ajout d'une colonne 'heure'
df_eco_2_cumuls_delta['heure'] = df_eco_2_cumuls_delta['date_heure'].dt.strftime('%H:%M')
df_eco_2_cumuls_delta['heure'].unique()
df_eco_2_cumuls_delta.info()

# Filtrage du dataframe sur les années 2013 à 2022

df_eco_2_cumuls_delta= df_eco_2_cumuls_delta[(df_eco_2_cumuls_delta['annee']>=2013) & (df_eco_2_cumuls_delta['annee']<=2022)]

# =======================================================================================
# Importation du fichier des relevés de températures et de vitesse du vent

df_temperatures = pd.read_csv("df_meteo_interpolated.csv", sep=',', encoding='latin-1')
df_temperatures.head()

df_temperatures.info()

# Conversion des dates 
df_temperatures['date_heure'] = pd.to_datetime(df_temperatures['date_heure'])

# Renommage des colonnes dans 'df_temperatures' pour correspondre aux noms dans 'df_eco_2_cumuls_delta'
df_temperatures.rename(columns={ 'Vitesse du vent moyen 10 mn': 'vent_10m_vitesse_moyen'}, inplace=True)
df_temperatures.info()

#========================FUSION =============================

# Fusion avec le DataFrame des températures
df_conso_temp = pd.merge(
    df_eco_2_cumuls_delta,
    df_temperatures[['date_heure','region','vent_10m_vitesse_moyen', 'temperature_C', 'temperature_ressentie']],
    on=['region', 'date_heure'],
    how='left')
df_conso_temp.info()

# Recherche de valeurs manquantes


df_conso_temp.isna().sum()

df_conso_temp_na = df_conso_temp[df_conso_temp.isna().any(axis=1)]

# Les valeurs manquantes correspondent aux relevés du 2022-12-31 23:30:00; elles peuvent être supprimmées sans nuire à l'étude
df_conso_temp.dropna(inplace = True)
df_conso_temp.isna().sum()

# Liste des colonnes de production pour lesquelles effectuer les calculs
colonnes_production = df_eco_2_cumuls_delta.columns[3:16]

# Affichage des informations pour vérification
df_conso_temp.info()

# Suppression des lignes où les températures et vitesse du vent sont des NaN
df_conso_temp = df_conso_temp.dropna(subset=['vent_10m_vitesse_moyen', 'temperature_C', 'temperature_ressentie'])

# ==================== Ajout des colonnes 'delta_consommation' et 'delta_TMoy'

# Calcul de la différence de consommation cumulée par rapport à la ligne précédente pour chaque région
df_conso_temp['delta_consommation'] = df_conso_temp.groupby('region')['consommation'].diff()

# Calcul de la différence de température réelle moyenne par rapport à la ligne précédente pour chaque région
df_conso_temp['delta_TMoy'] = df_conso_temp.groupby('region')['temperature_C'].diff()

# Vérification des résultats finaux
df_conso_temp.info()

# Ajout d'une colonne des variations de consommations par variation de la température réelle moyenne
# =====> Sensibilité de Consommation Électrique à la Température (SCET)

df_conso_temp['SCET'] = np.where(
    df_conso_temp['delta_TMoy'] == 0, np.nan,
    df_conso_temp['delta_consommation'] / df_conso_temp['delta_TMoy'] * 1_000_000)
# Remplacement des NaN par les 2 voisins les plus proches
df_conso_temp['SCET'] = df_conso_temp['SCET'].interpolate()

# Ajout d'une colonne de l'élasticité de la consommations électriques par rapport  aux variation de la température réelle moyenne
# =====> Elasticité Consommation Électrique/ Température (Elasticite_C_T)

df_conso_temp['Elasticite_C_T'] = df_conso_temp['delta_consommation']*df_conso_temp['temperature_C'] /df_conso_temp['consommation'].replace(0, np.nan)
# Remplacement des NaN par les 2 voisins les plus proches
df_conso_temp['Elasticite_C_T'] = df_conso_temp['Elasticite_C_T'].interpolate()
df_conso_temp.info()

# AJOUT DES POPULATIONS

df_population = pd.read_csv("recensement.csv", sep=';', encoding='latin-1')
df_population.head()

# Transformation de 'df_population' pour le passer en format long (melt)
df_population_long = df_population.melt(
    id_vars=['region'], 
    var_name='annee', 
    value_name='population'
)

# On s'assure que 'annee' dans les deux DataFrames est de type int pour la jointure
df_population_long['annee'] = df_population_long['annee'].astype(int)
df_conso_temp['annee'] = df_conso_temp['annee'].astype(int)

# Jointure sur 'region' et 'annee'
df_conso_temp = pd.merge(df_conso_temp, df_population_long, on=['region', 'annee'], how='left')

# Ajout d'une colonne de la SCETH par habitant
# =====> Sensibilité de Consommation Électrique à la Température (SCETH)

# Calcul de SCETH avec gestion des zéros
df_conso_temp['SCETH'] = np.where((df_conso_temp['delta_TMoy'] == 0) | (df_conso_temp['population'] == 0),np.nan,
    df_conso_temp['delta_consommation'] / df_conso_temp['delta_TMoy'] / df_conso_temp['population'] * 1_000_000)

# Remplacement des NaN par les 2 voisins les plus proches
df_conso_temp['SCETH'] = df_conso_temp['SCETH'].interpolate()

# Reccherche de valeurs manquantes
df_conso_temp.isna().sum()

# Suppression des lignes où les températures sont des NaN (extrémités des données)
df_conso_temp = df_conso_temp.dropna(subset=['SCET','Elasticite_C_T', 'SCETH'])
df_conso_temp.info()

df_conso_temp.isna().sum()

# AJOUT DES DONNEES SUR LA VITESSE DU VENT A 100M ET DES RAYONNEMENTS SOLAIRES

df_vent_100m_solaire_30min = pd.read_csv("df_vent_100m_solaire_30min.csv", sep=';', encoding='latin-1')
df_vent_100m_solaire_30min.head()

# Conversion des dates 
df_vent_100m_solaire_30min['date_heure'] = pd.to_datetime(df_vent_100m_solaire_30min['date_heure'])

# Fusion
df_conso_meteo = pd.merge(
    df_conso_temp,
    df_vent_100m_solaire_30min[['date_heure','region','vitesse_vent_100m_m_par_s', 'rayonnement_solaire_global_W_par_m2']],
    on=['region', 'date_heure'],
    how='left')
df_conso_meteo.info()

df_conso_meteo.isna().sum()

# AJOUT D'UNE COLONNE DU CUBE DE LA VITESSE DU VENT ET CALCUL DU COEFFICIENT EOLIEN REGIONAL POUR ANALYSE ULTERIEURE DE L'ENERGIE EOLIENNE

df_conso_meteo['vitesse_vent_100m_cube'] = df_conso_meteo['vitesse_vent_100m_m_par_s'] ** 3

df_conso_meteo['K_eolien_regional'] = np.where(
    df_conso_meteo['vitesse_vent_100m_cube'] != 0,
    (df_conso_meteo['eolien'] / df_conso_meteo['vitesse_vent_100m_cube']) * 1_000_000, 0)

# AJOUT D'UNE COLONNE DE CALCUL DU COEFFICIENT SOLAIRE REGIONAL POUR ANALYSE ULTERIEURE DE L'ENERGIE SOLAIRE

df_conso_meteo['K_solaire_regional'] = np.where(
    df_conso_meteo['rayonnement_solaire_global_W_par_m2'] != 0,
    (df_conso_meteo['solaire'] / df_conso_meteo['rayonnement_solaire_global_W_par_m2']) * 1_000_000,0)

print(df_conso_meteo.columns)

# Organisation des colonnes

ordre_desire =['reference','region', 'population','date_heure', 'date','annee', 'annee_mois', 'mois', 'mois_nom',
       'semaine', 'jour', 'jour_nom', 'jour_numero', 'heure',
       'jour_fractionnel','consommation', 'thermique',
       'nucleaire', 'eolien', 'solaire', 'hydraulique', 'bioenergies',
       'pompage', 'ech_physiques', 'export', 'import', 'stockage_batterie',
       'destockage_batterie', 'tco_6f', 'consommation_isol', 'tch_6f',
       'tch_consommation', 'tch_consommation_isol', 'delta_consommation','temperature_C', 'temperature_ressentie', 'delta_TMoy', 'SCET',
      'Elasticite_C_T', 'SCETH',
        'vent_10m_vitesse_moyen', 'vitesse_vent_100m_m_par_s','vitesse_vent_100m_cube','K_eolien_regional',
       'rayonnement_solaire_global_W_par_m2','K_solaire_regional']
df_conso_meteo = df_conso_meteo[ordre_desire]
df_conso_meteo.info()

df_conso_meteo.isna().sum()

# Sauvegarde des résultats
df_conso_meteo.to_csv("df_conso_meteo.csv", sep=';',index=False)
