==> 5-Fusion_eco2_capM_Ind_proc.ipynb <==
"""
 - Fusion des fichiers des capacités et des indisponibilités de production générés par les codes précédents (rub.3 et 4)
 - Création de colonnes calculées des puissances disponibles et de capacités maximales par filière
Created on 2024-06-15

@author: Thierry ALLEM
"""

import numpy as np
import pandas as pd
import seaborn as sns
import datetime
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import holidays

import warnings
# Pour ignorer tous les warnings
warnings.filterwarnings('ignore')

# ********************************************  FUSION  DES DONNEES ECO2MIX - CAPACITES MAXIMALES DE PRODUCTION ET INDISPONIBILITES DE PRODUCTION *********************************************************************

# Importation du fichier eco2mix_regional_def préparé, incluant les TCO
df_eco2_cap = pd.read_csv('df_eco2_cleaned_tco_tch.csv',sep = ';', encoding='latin-1')

df_eco2_cap.info()

df_eco2_cap_copy = df_eco2_cap.copy()

df_eco2_cap.head()

# Extraction des capacités maximales de production par région, filière et date effectuée dans code distinct 'capacites_productions' (copie du code c-dessous)

# Importation du fichier après extraction et calcul depuis le Registre des Installations

df_result_cap_prod = pd.read_csv('df_result_cap_prod_2013_2022.csv', sep=';')

df_result_cap_prod.head()

# Conversion de la colonne 'date' de df_result_cap_prod en datetime
df_result_cap_prod['date'] = pd.to_datetime(df_result_cap_prod['date'])
df_eco2_cap['date_heure'] = pd.to_datetime(df_eco2_cap['date_heure'])

# Pivot afin d'obtenir une colonne pour chaque filière de production
pivot_df = df_result_cap_prod.pivot_table(index=['region', 'date'],
                                          columns='filiere_production',
                                          values='puis_max_installee').reset_index()

# Renommage des colonnes avant intégration dans df_eco2_cap
column_mapping = {
    'bioenergies': 'cap_prod_max_bioenergies',
    'eolien': 'cap_prod_max_eolien',
    'hydraulique': 'cap_prod_max_hydraulique',
    'nucleaire': 'cap_prod_max_nucleaire',
    'solaire': 'cap_prod_max_solaire',
    'thermique': 'cap_prod_max_thermique'
}

# Ajout de 'region' et 'date' au mapping si elles sont présentes
new_columns = ['region', 'date'] + [column_mapping.get(col, col) for col in pivot_df.columns if col not in ['region', 'date']]
pivot_df.columns = new_columns

# Jointure de pivot_df avec df_eco2_cap_ind sur les colonnes 'region' et 'date_heure'
df_eco2_cap = df_eco2_cap.merge(pivot_df, left_on=['region', 'date_heure'], right_on=['region', 'date'], how='left')

# Vérification que la colonne 'date' existe avant de la supprimer
if 'date' in df_eco2_cap.columns:
    df_eco2_cap.drop(columns=['date'], inplace=True)


# Vérification du résultat
df_eco2_cap.head()

# **************************************************** AJOUT DES RESULTATS DES EXTRACTIONS DES INDISPONIBILITES DES MOYENS DE PRODUCTIONS *****************
# Extraction des indisponibilités des moyens de production par région, filière et date effectuée dans code distinct 'indisponibilites' (copie du code c-dessous)
# Importation du fichier après extraction et calcul depuis le Registre des Installations

# Importation du fichier des capacités de productions totales
df_result_ind_prod = pd.read_csv('df_result_ind_prod_prefusion.csv', sep=';')

# Conversion de la colonne 'date' de df_result_cap_prod en datetime
df_result_ind_prod['date'] = pd.to_datetime(df_result_ind_prod['date'])

df_result_ind_prod.head()

df_result_ind_prod['filiere'].unique()

# Pivot afin d'obtenir une colonne pour chaque filière de production
pivot_df2 = df_result_ind_prod.pivot_table(index=['region', 'date'],
                                          columns='filiere',
                                          values='capacite_indisponible').reset_index()

# Renommage des colonnes avant intégration dans df_eco2_cap_ind
column_mapping2 = {'bioenergies': 'ind_prod_bioenergies','hydraulique': 'ind_prod_hydraulique',
    'nucleaire': 'ind_prod_nucleaire','thermique': 'ind_prod_thermique'}

# Ajout de 'region' et 'date' au mapping si elles sont présentes
new_columns2 = ['region', 'date'] + [column_mapping2.get(col, col) for col in pivot_df2.columns if col not in ['region', 'date']]
pivot_df2.columns = new_columns2

# Jointure de pivot_df2 avec df_eco2_cap_ind sur les colonnes 'region' et 'date_heure'
df_eco2_cap_ind = df_eco2_cap.merge(pivot_df2, left_on=['region', 'date_heure'], right_on=['region', 'date'], how='left')

# Vérification que la colonne 'date' existe avant de la supprimer
if 'date' in df_eco2_cap_ind.columns:
    df_eco2_cap_ind.drop(columns=['date'], inplace=True)


# Remplacement des valeurs manquantes par la valeur 0
df_eco2_cap_ind.fillna(0, inplace = True)
print(df_eco2_cap_ind.isna().sum())

# Arrondi des valeurs des capacités et indisponibilités à 2 décimales
df_eco2_cap_ind.iloc[:, 9:19] = df_eco2_cap_ind.iloc[:, 9:19].round(2)

# AJOUT DE COLONNES DE DONNEES TEMPORELLES

# Ajout de la colonne 'annee'
df_eco2_cap_ind['annee'] = df_eco2_cap_ind['date_heure'].dt.year

# Ajout de la colonne 'mois_nom'
df_eco2_cap_ind['mois_nom'] = df_eco2_cap_ind['date_heure'].dt.strftime('%b')

# Ajout de la colonne 'mois'
df_eco2_cap_ind['mois'] = df_eco2_cap_ind['date_heure'].dt.month

# Ajout de la colonne 'jour_nom'
df_eco2_cap_ind['jour_nom'] = df_eco2_cap_ind['date_heure'].dt.strftime('%a')

# Ajout de la colonne 'jour'
df_eco2_cap_ind['jour'] = df_eco2_cap_ind['date_heure'].dt.day

# Ajout de la colonne 'jour_numero'
df_eco2_cap_ind['jour_numero'] = df_eco2_cap_ind['date_heure'].dt.strftime('%j')

# Ajout de la colonne 'semaine'
df_eco2_cap_ind['semaine'] = df_eco2_cap_ind['date_heure'].dt.strftime('%U')

# Ajout de la colonne 'heure'
# df_eco2_cap_ind['heure'] = df_eco2_cap_ind['date_heure'].dt.strftime('%X')

# Ajout de la colonne 'jour_ferie'
jours_feries_fr = holidays.France(years=range(2013, 2023))
df_eco2_cap_ind['jour_ferie'] = df_eco2_cap_ind['date_heure'].dt.strftime('%Y-%m-%d').isin(jours_feries_fr)

#  ************ AJOUT D'UNE COLONNE PORTANT LES NOMS DE REGIONS ABREGES SELON LA NORME ISO 3166-2 ****************************************************
        # Insertion d'une nouvelle colonne 'region_abr' à l'index 1
df_eco2_cap_ind.insert(1, 'region_abr', df_eco2_cap_ind['region'])

        # Affectation des valeurs de 'region_ncc' à 'region_abr'
df_eco2_cap_ind['region_abr'] = df_eco2_cap_ind['region']

        # Renommage des valeurs de 'region_abr'
dictionnaire_region_abr= {"PROVENCE ALPES COTE D AZUR" : "FR-PAC",
                      "BOURGOGNE FRANCHE COMTE" : "FR-BFC",
                      "BRETAGNE" : "FR-BRE",
                      "NORMANDIE" : "FR-NOR",
                      "NOUVELLE AQUITAINE" : "FR-NAQ",
                      "HAUTS DE FRANCE" : "FR-HDF",
                      "ILE DE FRANCE" : "FR-IDF",
                      "AUVERGNE RHONE ALPES":"FR-ARA",
                      "GRAND EST":"FR-GES",
                      "CENTRE VAL DE LOIRE":"FR-CVL",
                      "OCCITANIE":"FR-OCC",
                      "PAYS DE LA LOIRE":"FR-PDL"}
df_eco2_cap_ind['region_abr']=df_eco2_cap_ind['region_abr'].replace(dictionnaire_region_abr)
print("Noms abrégés des régions : ", df_eco2_cap_ind['region_abr'].unique())

# Renommage de colonnes
df_eco2_cap_ind = df_eco2_cap_ind.rename(columns={'eolien_x':'eolien', 'solaire_x' : 'solaire'})

print("Noms abrégés des régions après renommage: ", df_eco2_cap_ind['region_abr'].unique())

# **************************************************** AJOUT DES CALCULS  DES PUISSANCES DISPONIBLES PAR FILIERE *****************
# Ajout des capacités de production disponibles en tenant compte des indisponibilités des moyens de production


thermique_p_disp = df_eco2_cap_ind['cap_prod_max_thermique'] - df_eco2_cap_ind['ind_prod_thermique']
nucleaire_p_disp = df_eco2_cap_ind['cap_prod_max_nucleaire'] - df_eco2_cap_ind['ind_prod_nucleaire']
hydraulique_p_disp = df_eco2_cap_ind['cap_prod_max_hydraulique'] - df_eco2_cap_ind['ind_prod_hydraulique']
bioenergies_p_disp = df_eco2_cap_ind['cap_prod_max_bioenergies'] - df_eco2_cap_ind['ind_prod_bioenergies']
df_eco2_cap_ind['thermique_p_disp'] = np.around(thermique_p_disp,0)
df_eco2_cap_ind['nucleaire_p_disp'] = np.around(nucleaire_p_disp,0)
df_eco2_cap_ind['hydraulique_p_disp'] = np.around(hydraulique_p_disp,0)
df_eco2_cap_ind['bioenergies_p_disp'] = np.around(bioenergies_p_disp,0)
    
#   A noter que les registres des indisponibilités des filières 'eolien' et solaire ne sont pas disponibles auprès de RTE
#       Renommage des colonnes cap_prod_max_eolien et cap_prod_max_eolien (à défaut de données, on suppose les indisponibilités nulles)

dictionnaire_cap_prod_disp= {'cap_prod_max_eolien':'eolien_p_disp', 'cap_prod_max_solaire':'solaire_p_disp'} 
df_eco2_cap_ind=df_eco2_cap_ind.rename(dictionnaire_cap_prod_disp, axis = 1)

# CALCUL DES TCH (Taux de Charge des filières)  PAR REGION ET PAR FILIERE - Les batteries de stockage et les échanges physiques (import/export)
#ne sont pas inclus dans les TCH

tch_thermique = np.where(df_eco2_cap_ind['thermique_p_disp'] > 0, df_eco2_cap_ind['thermique']/df_eco2_cap_ind['thermique_p_disp']*100, 0)
tch_nucleaire= np.where(df_eco2_cap_ind['nucleaire_p_disp'] > 0, df_eco2_cap_ind['nucleaire']/df_eco2_cap_ind['nucleaire_p_disp']*100, 0)
tch_eolien = np.where(df_eco2_cap_ind['eolien_p_disp'] > 0, df_eco2_cap_ind['eolien']/df_eco2_cap_ind['eolien_p_disp']*100, 0)
tch_solaire= np.where(df_eco2_cap_ind['solaire_p_disp'] > 0, df_eco2_cap_ind['solaire']/df_eco2_cap_ind['solaire_p_disp']*100, 0)

tch_hydraulique=np.where(df_eco2_cap_ind['hydraulique_p_disp']>0, df_eco2_cap_ind['hydraulique']/df_eco2_cap_ind['hydraulique_p_disp']*100, 0)
tch_bioenergies=np.where(df_eco2_cap_ind['bioenergies_p_disp']>0, df_eco2_cap_ind['bioenergies']/df_eco2_cap_ind['bioenergies_p_disp']*100, 0)

# Calcul des TCH totaux par région (regroupement de toutes les filières de production)
region_p_disp = (df_eco2_cap_ind['thermique_p_disp'] + df_eco2_cap_ind['nucleaire_p_disp'] + df_eco2_cap_ind['eolien_p_disp']+ 
                 df_eco2_cap_ind['solaire_p_disp'] + df_eco2_cap_ind['hydraulique_p_disp']+df_eco2_cap_ind['bioenergies_p_disp'])
df_eco2_cap_ind['region_p_disp'] = round(region_p_disp, 0)

region_p_reelle = (df_eco2_cap_ind['thermique'] + df_eco2_cap_ind['nucleaire'] + df_eco2_cap_ind['eolien'] + df_eco2_cap_ind['solaire']
                       + df_eco2_cap_ind['hydraulique'] + df_eco2_cap_ind['bioenergies'])
df_eco2_cap_ind['region_p_reelle'] = round(region_p_reelle, 0) 

tch_region = np.where(df_eco2_cap_ind['region_p_disp'] > 0, df_eco2_cap_ind['region_p_reelle'] / df_eco2_cap_ind['region_p_disp']*100, 0)

df_eco2_cap_ind['tch_thermique'] = np.around(tch_thermique,2)
df_eco2_cap_ind['tch_nucleaire'] = np.around(tch_nucleaire,2)
df_eco2_cap_ind['tch_eolien'] = np.around(tch_eolien,2)
df_eco2_cap_ind['tch_solaire'] = np.around(tch_solaire,2)
df_eco2_cap_ind['tch_hydraulique'] = np.around(tch_hydraulique,2)
df_eco2_cap_ind['tch_bioenergies'] = np.around(tch_bioenergies,2)
df_eco2_cap_ind['tch_region'] = np.around(tch_region,2)

# Suppression de la colonne 'Unnamed: 0'
# col_to_drop = ['Unnamed: 0','nature','date_heure_utc','date_y']
col_to_drop = ['Unnamed: 0','date_y']
df_eco2_cap_ind = df_eco2_cap_ind.drop(col_to_drop, axis=1)

# Renommage de colonnes
df_eco2_cap_ind = df_eco2_cap_ind.rename(columns={'date_x':'date'})
df_eco2_cap_ind = df_eco2_cap_ind.rename(columns={'stockage_batterie_x':'stockage_batterie'})
df_eco2_cap_ind = df_eco2_cap_ind.rename(columns={'date_ym':'annee_mois'})

# ********** AJ0UT DE LA CAPACITE MAXIMALE DE PRODUCTION DE CHAQUE REGION SANS PRISE EN COMPTE DES INDISPONIBILITES *****************

region_p_max = (df_eco2_cap_ind['cap_prod_max_thermique'] + df_eco2_cap_ind['cap_prod_max_nucleaire'] + df_eco2_cap_ind['eolien_p_disp']+ 
                 df_eco2_cap_ind['solaire_p_disp'] + df_eco2_cap_ind['cap_prod_max_hydraulique']+df_eco2_cap_ind['cap_prod_max_bioenergies'])
df_eco2_cap_ind['region_p_max'] = np.around(region_p_max,2)


# Réorganisation de l'ordre des colonnes

ordre_desire2 = ['region_abr', 'region', 'date_heure', 'date', 'heure', 'annee_mois', 'annee', 'mois', 'mois_nom', 
                 'jour', 'jour_nom', 'jour_numero','semaine', 'jour_ferie',
                 'consommation', 'thermique', 'nucleaire', 'eolien', 'solaire', 'hydraulique', 'bioenergies','pompage', 'ech_physiques','export', 'import',
                 'stockage_batterie', 'destockage_batterie',
                 'tch_thermique_rte', 'tch_nucleaire_rte','tch_eolien_rte','tch_solaire_rte','tch_hydraulique_rte','tch_bioenergies_rte',
                 'cap_prod_max_thermique', 'ind_prod_thermique', 'thermique_p_disp', 'cap_prod_max_nucleaire', 'ind_prod_nucleaire', 'nucleaire_p_disp',
                 'eolien_p_disp','solaire_p_disp', 'cap_prod_max_hydraulique', 'ind_prod_hydraulique', 'hydraulique_p_disp', 
                 'cap_prod_max_bioenergies','ind_prod_bioenergies','bioenergies_p_disp', 'region_p_max',
                 'tco_thermique', 'tco_nucleaire', 'tco_eolien','tco_solaire', 'tco_hydraulique', 'tco_pompage', 'tco_bioenergies',
                 'tco_export','tco_import', 'tco_stockage_batterie', 'tco_destockage_batterie']



df_eco2_cap_ind = df_eco2_cap_ind[ordre_desire2]


df_eco2_cap_ind.info()

print ("Valeurs manquantes sur dataframe final : ",df_eco2_cap_ind.isnull().sum().sum())

# Arrondi des valeurs des capacités et indisponibilités à 0 décimale
df_eco2_cap_ind.iloc[:, 27:43] = df_eco2_cap_ind.iloc[:, 27:43].round(0)

df_eco2_cap_ind.info()

# Export du dataFrame df_eco2_cap_ind au format CSV
df_eco2_cap_ind.to_csv('df_eco2_cap_ind_fusion.csv', sep=';', encoding='latin-1')

# Export des 100 premières lignes dans un nouveau fichier CSV pour aperçu
df_eco2_cap_ind_head = df_eco2_cap_ind.head(100)
df_eco2_cap_ind_head.to_csv('df_eco2_cap_ind_first_100_lines.csv', sep=';', encoding='latin-1', index=False)
